[Passed] 가장 짧은 문자열 (1ms, 0MB)

+++ Codetree_V2/trail1/가장 짧은 문자열/shortest-string.cpp
@@ -5,8 +5,8 @@ int main() {
     string str[4];
     cin >> str[1] >> str[2] >> str[3];
     
-    int minN = min({ str[1].length(), str[2].length(), str[3].length() });
-    int maxN = max({ str[1].length(), str[2].length(), str[3].length() });
+    int minN = min({ (int)str[1].length(), (int)str[2].length(), (int)str[3].length() });
+    int maxN = max({ (int)str[1].length(), (int)str[2].length(), (int)str[3].length() });
 
     cout << maxN-minN;
     return 0;
